requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
SQLAlchemy==2.0.32
pymysql==1.1.1
alembic==1.13.2
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
openpyxl==3.1.5
python-dotenv==1.0.1